reuse the same code for electron and browser
by extracting the common part to a cordova new module
