--- bindings/wallet-cordova/src/browser/Wallet.js
@@ -1,1 +1,0 @@
-// TODO
